pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.9",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "docxtpl>=0.20.2",
    "pillow>=12.0.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "weasyprint>=67.0",
]
